change swatch and checkerboard sizes to avoid vertical color bleed
- text position and spacing between swatches is also modified to rebalance the final display elements

.# Discussion

For some fonts, notably many of the "[nerd fonts](https://www.nerdfonts.com)", the
block characters used as "pixels" are not the pixel-perfect character width. So, for
characters which are a mixture of colors, some color bleed can occur. This is most
noticeable between the solid color swatch and the surrounding checkerboard.

Modifying the color swatch size to be vertically offset by some multiple of a full
character (or, two "pixels") from the top and bottom of the checkerboard avoids this
bleed effect. Since the swatch is centered on the checkerboard, its total size
difference, in "pixels", from the checkerboard size must be a multiple of
four (eg, 0, 4, 8, ...) to avoid the bleed effect.