change ~ add vertical padding between printed swatches
- a leading vertical pad is printed for better symmetry

.# Discussion

For output visual symmetry, a leading line is printed as padding. The strategy used
requires a bit of state to track the first/later swatch output. The implementation uses
an `atomic` which may be a bit over-the-top for this use case, but I don't see another
"safe" implementation at this point.

If further output might be changed based upon swatch count, an `Atomic::Usize` could be
used instead (see surrounding comments in code for the implementation).