Fix lines in kitty terminal with text_fg_override_threshold set
The kitty terminal has a new setting text_fg_override_threshold that checks
the luminosity difference between the text and background and if it is above
the set percentage turns the text black or white to get the best possible
visibility.  pastel currently uses a unicode half block character and sets
both forground and background to the same color in the color patch.  This
patch changes this to be a space with both forground and background colors set
to the same color (only becuase I don't know enough rust to figure out how to
just set background color :/).

Possibly this could cause trouble with some background transparency settings,
although if so the current situation might as well to a lesser extent.  I'm
not sure how to make using text color work reliably with
text_fg_override_threshold so I think a command line option to use one or the
other would be needed there unless another drawing option is used.

There are two other ways to draw on kitty that could be considered.  The
graphics protocol allow arbitrary images to be displayed and is supported by a
few other terminals.  kitty also supports using colors with DECCARA, although
I don't know if any other terminals support this.  In any case, with this
patch it works for me and I checked that it still works on alacritty as well.